Correct unusued name assignment
I should not ignore that assign_unique_name parameter!
